+++ spec/attribute_spec.rb
@@ -157,7 +157,7 @@ describe Attributor::Attribute do
       end
 
       it 'passes any given parent through to the example proc' do
-        subject.example(nil, some_object).should == 'ok'
+        subject.example(nil, parent: some_object).should == 'ok'
       end
 
     end
@@ -413,25 +413,15 @@ describe Attributor::Attribute do
         let(:chicken) { Chicken.example }
         let(:type_attributes) { type.definition.attributes }
 
-
-        let(:email_validation_response) { [] }
-        let(:age_validation_response) { [] }
-
-        before do
-          type_attributes[:email].should_receive(:validate).
-            with(chicken.email, :email).and_return(email_validation_response)
-          type_attributes[:age].should_receive(:validate).
-            with(chicken.age, :age).and_return(age_validation_response)
-        end
-
         it 'validates sub-attributes' do
           errors = attribute.validate(chicken)
           errors.should be_empty
         end
 
         context 'with a failing validation' do
-          let(:email_validation_response) { ["Chicken has invalid email"] }
-          let(:age_validation_response) { ["Chicken is too old"] }
+          subject(:chicken) { Chicken.example(age: 150, email: "foo") }
+          let(:email_validation_response) { ["email value does not match regexp (/@/)"] }
+          let(:age_validation_response) { ["age value is larger than the allowed max (120)"] }
 
           it 'collects sub-attribute validation errors' do
             errors = attribute.validate(chicken)